requirements.txt
moviepy==1.0.3
mutagen==1.45.1
playwright==1.22.0
praw==7.6.0
rich==12.4.4
yt_dlp==2022.5.18
google-cloud-texttospeech
pydantic[dotenv]
typer
sqlmodel
SQLAlchemy==1.4.35
pyinstrument
python-slugify
sox==1.4.1
requests